sentiment-service-backend/requirements.txt
feedparser~=6.0.11
Flask~=3.1.1
requests~=2.32.4
bs4~=0.0.2
beautifulsoup4~=4.13.4
google-generativeai~=0.8.5
python-dotenv~=1.1.1
pymongo~=4.13.2
apscheduler~=3.11.0
flask-cors~=6.0.1
pytimeparse~=1.1.8
python-dateutil~=2.9.0.post0
transformers~=4.55.4
torch~=2.8.0